chore: move enum to utils
